refactor(svelte): split focus-state tests by decision group
Extract the 807-line legend focus-state harness monolith into a shared
harness plus handlers, reconcile, and integration suites so lifecycle
seams match inspection/interval state locality.
